Backend/requirements/development.txt
-r base.txt

# ── Debug toolbar ─────────────────────────────────────────────────────────────
django-debug-toolbar==4.4.6

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.1.0

# ── Code quality ──────────────────────────────────────────────────────────────
ruff==0.9.2
mypy==1.14.1
django-stubs==5.1.1
djangorestframework-stubs==3.15.1

# ── Shell & utilities ─────────────────────────────────────────────────────────
ipython==8.31.0
django-extensions==3.2.3
